+++ PathWeaver/pathweaver.json
@@ -1,5 +1,5 @@
 {
-  "lengthUnit": "Inch",
+  "lengthUnit": "Foot",
   "exportUnit": "Same as Project",
   "maxVelocity": 132.0,
   "maxAcceleration": 90.0,
